requirements.txt
python-magic==0.4.15

